update by mac os , complete Gp

+++ LeetCode/LetterCombinations.cpp
@@ -31,7 +31,7 @@ public:
             vcs.push_back(digits_map[digits[i] - 48]);
         }
 
-        Comb(0, s, vcs, "", digits.length());
+        Comb(0, s, vcs, "", (int) digits.length());
 
         return s;
     }
